Refactor: fixed broken tests
