+++ packages/protocol/versioning.go
@@ -1,5 +1,7 @@
 package protocol
 
-import "github.com/iotaledger/goshimmer/packages/protocol/database"
+import (
+	"github.com/iotaledger/goshimmer/packages/core/database"
+)
 
 const DatabaseVersion database.Version = 1
