+++ plugins/metrics/epochcommittment.go
@@ -7,14 +7,15 @@ import (
 
 	"github.com/iotaledger/goshimmer/packages/core/commitment"
 	"github.com/iotaledger/goshimmer/packages/protocol/chainmanager"
+	"github.com/iotaledger/goshimmer/packages/protocol/engine/notarization"
 )
 
 var (
-	lastCommittedEpoch      = chainmanager.NewCommitment(commitment.NewMerkleRoot([]byte{}))
+	lastCommittedEpoch      = chainmanager.NewCommitment(commitment.NewID(0, []byte{}))
 	lastCommittedEpochMutex sync.RWMutex
 )
 
-var onEpochCommitted = event.NewClosure(func(event *notarizationold.EpochCommittableEvent) {
+var onEpochCommitted = event.NewClosure(func(event *notarization.EpochCommittableEvent) {
 	lastCommittedEpochMutex.Lock()
 	defer lastCommittedEpochMutex.Unlock()
 	lastCommittedEpoch = event.ECRecord
