+++ plugins/peer/plugin.go
@@ -22,8 +22,8 @@ import (
 	"github.com/iotaledger/hive.go/core/kvstore"
 	"github.com/iotaledger/hive.go/core/node"
 
+	"github.com/iotaledger/goshimmer/packages/core/database"
 	"github.com/iotaledger/goshimmer/packages/core/shutdown"
-	databasePkg "github.com/iotaledger/goshimmer/packages/protocol/database"
 	"github.com/iotaledger/goshimmer/plugins/protocol"
 )
 
@@ -167,12 +167,12 @@ func initPeerDB() (*peer.DB, kvstore.KVStore, bool, error) {
 		return nil, nil, false, err
 	}
 
-	db, err := databasePkg.NewDB(Parameters.PeerDBDirectory)
+	db, err := database.NewDB(Parameters.PeerDBDirectory)
 	if err != nil {
 		return nil, nil, false, fmt.Errorf("error creating peer database: %s", err)
 	}
 
-	peerDBKVStore, err := db.NewStore().WithRealm([]byte{databasePkg.PrefixPeer})
+	peerDBKVStore, err := db.NewStore().WithRealm([]byte{database.PrefixPeer})
 	if err != nil {
 		return nil, nil, false, fmt.Errorf("error creating peer store: %w", err)
 	}
